app/requirements.txt
pymongo
pandas
python-dotenv
watchdog